% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{run_cli}
\alias{run_cli}
\title{Run hotwater from the command line}
\usage{
hotwater [args]
}
\arguments{
\item{-f}{plumber file}

\item{-d}{extra directories}

\item{-p}{plumber port}

\item{-h}{show help}

\item{--host}{plumber host}
}
\description{
Following \code{\link[=install_hotwater]{install_hotwater()}}, the \code{hotwater} command can be used
to run a hotwater engine straight from the terminal. See
\code{\link[=run]{run()}} for further details on default values.

\code{hotwater -v} will provide the current version of hotwater.
\code{hotwater -h} will provide help text.
}
\examples{
```sh
hotwater -f path/to/app.R -p 9999
```
}
\seealso{
\link{run}
}
