% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwater-package.R
\docType{package}
\name{hotwater-package}
\alias{hotwater}
\alias{hotwater-package}
\title{hotwater: Live Reload for Plumber APIs}
\description{
Enhances development for plumber APIs by enabling live reloading. Monitors API files for changes and automatically refreshes the server and connected web clients, allowing for faster API iteration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElianHugh/hotwater}
  \item \url{https://elianhugh.github.io/hotwater/}
  \item Report bugs at \url{https://github.com/ElianHugh/hotwater/issues}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
