% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{install_hotwater}
\alias{install_hotwater}
\title{WORK IN PROGRESS}
\usage{
install_hotwater(install_folder = "~/.local/bin/")
}
\arguments{
\item{install_folder}{folder (in PATH) to install hotwater}
}
\description{
WORK IN PROGRESS
}
