% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Start hotwater engine}
\usage{
run(path, dirs = NULL, port = NULL, host = NULL, ignore = NULL)
}
\arguments{
\item{path}{path to plumber file}

\item{dirs}{extra directories to watch}

\item{port}{port to launch API on, defaults to \code{httpuv::randomPort()}}

\item{host}{host to launch API on, defaults to "127.0.0.1"}

\item{ignore}{vector of files or file extensions to ignore (globs)}
}
\description{
Start a hotwater engine, which launches a plumber API that is restarted whenever
the plumber API's folder is modified.

Extra directories can be specified to refresh the API when directories other than the plumber folder are modified.

If a plumber endpoint returns an html response, when hotwater refreshes the API, hotwater will also order
a refresh of any webpage that is using the API.
}
\examples{
if (interactive()) {
 hotwater::run(system.file("examples", "plumber.R", package = "hotwater"))
}

}
