% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{uninstall_hotwater}
\alias{uninstall_hotwater}
\title{WORK IN PROGRESS}
\usage{
uninstall_hotwater(install_folder = "~/.local/bin/")
}
\arguments{
\item{install_folder}{folder (in PATH) to uninstall hotwater}
}
\description{
WORK IN PROGRESS
}
