% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{uninstall_hotwater}
\alias{uninstall_hotwater}
\title{Uninstall global hotwater script}
\usage{
uninstall_hotwater(install_folder)
}
\arguments{
\item{install_folder}{folder to uninstall hotwater from.}
}
\description{
Uninstall global hotwater script
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 hotwater::uninstall_hotwater()
\dontshow{\}) # examplesIf}
}
\seealso{
\link{install_hotwater}
}
